+++ app.js
@@ -1,0 +1,14 @@
+require('dotenv').config()
+const express = require('express')
+
+
+const app = express()
+const PORT = process.env.PORT || 8080
+
+
+app.use(express.json())
+app.use(cors())
+app.use("/")
+
+app.listen(PORT, () => console.log(`Listening on ${PORT}`))
+
