+++ src/data/categories-data.json
@@ -1,0 +1,11 @@
+[
+    {
+        "categoryName": "Casa"
+    },
+    {
+        "categoryName": "Deportes"
+    },
+    {
+        "categoryName": "Viajes"
+    }
+]
