+++ src/data/tasks-data.json
@@ -1,0 +1,26 @@
+[
+    {
+        "name": "sexo",
+        "deadLine": "3/3/3",
+        "category": 2,
+        "completed": true
+    },
+    {
+        "name": "casa",
+        "deadLine": "213123",
+        "category": 0,
+        "completed": false
+    },
+    {
+        "name": "Correr",
+        "deadLine": "12/10/2033",
+        "category": 1,
+        "completed": false
+    },
+    {
+        "name": "asd",
+        "completed": false,
+        "deadLine": null,
+        "category": 2
+    }
+]
