--- tasksData.json
@@ -1,8 +1,0 @@
-[
-    {
-        "name": "Casa",
-        "completed": false,
-        "deadLine": "12-12-12",
-        "category": "Otros"
-    }
-]
